Refactor e2e_fees to use snapshots
